requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
bcrypt==4.0.1
pandas
numpy
scikit-learn
python-dotenv
